StreamingClientHandler interface
